Extract common functionality in Logback and Log4j 2 tests
