+++ log4j2-ecs-layout/src/test/java/co/elastic/logging/log4j2/Log4j2EcsLayoutTest.java
@@ -24,33 +24,22 @@
  */
 package co.elastic.logging.log4j2;
 
-import co.elastic.logging.AbstractEcsLoggingTest;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.TextNode;
 import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.Marker;
-import org.apache.logging.log4j.MarkerManager;
-import org.apache.logging.log4j.ThreadContext;
 import org.apache.logging.log4j.core.Appender;
 import org.apache.logging.log4j.core.BasicConfigurationFactory;
 import org.apache.logging.log4j.core.Logger;
 import org.apache.logging.log4j.core.LoggerContext;
 import org.apache.logging.log4j.core.config.ConfigurationFactory;
 import org.apache.logging.log4j.core.util.KeyValuePair;
-import org.apache.logging.log4j.message.StringMapMessage;
 import org.apache.logging.log4j.test.appender.ListAppender;
 import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.util.Map;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-class Log4j2EcsLayoutTest extends AbstractEcsLoggingTest {
+class Log4j2EcsLayoutTest extends AbstractLog4j2EcsLayoutTest {
 
     private static ConfigurationFactory configFactory = new BasicConfigurationFactory();
     private LoggerContext ctx = LoggerContext.getContext();
@@ -70,7 +59,7 @@ class Log4j2EcsLayoutTest extends AbstractEcsLoggingTest {
     }
 
     @BeforeEach
-    void setUp() throws Exception {
+    void setUp() {
         for (final Appender appender : root.getAppenders().values()) {
             root.removeAppender(appender);
         }
@@ -89,60 +78,6 @@ class Log4j2EcsLayoutTest extends AbstractEcsLoggingTest {
         root.addAppender(listAppender);
         root.setLevel(Level.DEBUG);
         ctx.getConfiguration().getProperties().put("node.id", "foo");
-    }
-
-    @AfterEach
-    void tearDown() throws Exception {
-        ThreadContext.clearAll();
-    }
-
-    @Test
-    void globalLabels() throws Exception {
-        putMdc("trace.id", "foo");
-        debug("test");
-        assertThat(getLastLogLine().get("cluster.uuid").textValue()).isEqualTo("9fe9134b-20b0-465e-acf9-8cc09ac9053b");
-        assertThat(getLastLogLine().get("node.id").textValue()).isEqualTo("foo");
-        assertThat(getLastLogLine().get("404")).isNull();
-    }
-
-    @Test
-    void testMarker() throws Exception {
-        Marker parent = MarkerManager.getMarker("parent");
-        Marker child = MarkerManager.getMarker("child").setParents(parent);
-        Marker grandchild = MarkerManager.getMarker("grandchild").setParents(child);
-        root.debug(grandchild, "test");
-
-        assertThat(getLastLogLine().get("tags")).contains(
-                TextNode.valueOf("parent"),
-                TextNode.valueOf("child"),
-                TextNode.valueOf("grandchild"));
-    }
-
-    @Test
-    void testMapMessage() throws Exception {
-        root.info(new StringMapMessage(Map.of("foo", "bar")));
-        assertThat(getLastLogLine().get("labels.foo").textValue()).isEqualTo("bar");
-    }
-
-    @Override
-    public void putMdc(String key, String value) {
-        ThreadContext.put(key, value);
-    }
-
-    @Override
-    public boolean putNdc(String message) {
-        ThreadContext.push(message);
-        return true;
-    }
-
-    @Override
-    public void debug(String message) {
-        root.debug(message);
-    }
-
-    @Override
-    public void error(String message, Throwable t) {
-        root.error(message, t);
     }
 
     @Override
